Moved legal to settings
