+++ apps/flutter_parent/lib/screens/settings/legal_screen.dart
