+++ apps/flutter_parent/test/screens/settings/legal_screen_test.dart
@@ -13,7 +13,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 import 'package:flutter_parent/l10n/app_localizations.dart';
 import 'package:flutter_parent/router/panda_router.dart';
-import 'package:flutter_parent/screens/help/legal_screen.dart';
+import 'package:flutter_parent/screens/settings/legal_screen.dart';
 import 'package:flutter_parent/utils/quick_nav.dart';
 import 'package:flutter_parent/utils/url_launcher.dart';
 import 'package:flutter_test/flutter_test.dart';
